Add files via upload
This project was written in python (using Syder). It shows the number of moves made by the player and should display the amount of time it took the player to complete the game (min:sec), the timer is supposed to restart when the player presses the reset button. Right now, however, the timer only shows seconds and does not reset, instead it shows the amount of time the program has been running.